Refactor: MainPage 내 Blog 폴더까지 절대경로 설정
